Rename class files from kebab-case to PascalCase
When file exports an object or a instance of a class, its name MUST be
in kebab-case.

When file exports a class, it MUST be in PascalCase.
