+++ api/tests/unit/infrastructure/caches/RedisCache_test.js
@@ -1,7 +1,7 @@
 const { expect, sinon } = require('../../../test-helper');
 const settings = require('../../../../lib/config');
 const Redlock = require('redlock');
-const RedisCache = require('../../../../lib/infrastructure/caches/redis-cache');
+const RedisCache = require('../../../../lib/infrastructure/caches/RedisCache');
 
 describe('Unit | Infrastructure | Cache | redis-cache', () => {
